requirements.txt
streamlit
tensorflow-cpu
numpy
matplotlib
pillow
opencv-python-headless